LayoutTests/fast/forms/validation-message-on-range-expected.txt

Check if the position of a validation message for a range control is good.


The top of the bubble should be on the bottom of the host node:
PASS bubbleRect.top is hostRect.bottom

PASS successfullyParsed is true

TEST COMPLETE
